Finished configuring project

+++ src/main/java/com/project/book_tables/DBs/Controllers/TablesController.java
@@ -24,22 +24,31 @@ public class TablesController {
         return tablesService.getAllTables();
     }
 
-    @PostMapping("/save_table")
-    public void createTable(@RequestParam String name) {
+    @PostMapping("/create_table")
+    public Tables createTable(@RequestParam String name) {
         Tables table = new Tables();
         table.setName(name);
-        tablesService.createTable(table);
+        return tablesService.createTable(table);
     }
 
     @DeleteMapping("/delete_table")
     public void deleteTable(@RequestParam Long id) {
+        if (tablesService.getTableById(id) == null) {
+            System.out.println("Не существует столика!");
+            return;
+        }
         tablesService.deleteTableById(id);
     }
 
     //id кого меняем, name что меняем
     @PostMapping("/update_table")
-    public void updateTable(@RequestParam Long id, @RequestParam String name)
-    {
-        tablesService.updateTable(id, name);
+    public Tables updateTable(@RequestParam Long id, @RequestParam String name) {
+        Tables table = tablesService.getTableById(id);
+        if (table == null) {
+            System.out.println("Не существует столика!");
+            return null;
+        }
+        table.setName(name);
+        return tablesService.updateTable(table);
     }
 }
